requirements.txt
numpy==1.24.4
streamlit==1.36.0
ultralytics==8.3.3
pillow==9.4.0